Delegate requirements.txt packages to setup.py
This moves the lists of required packages out of requirements.txt and in
to setup.py. This reduces duplication while maintaining flexibility and
staying true to the spirit of where what package requirements should be
expressed. Now, requirements.txt will install the armi package in
"editable" mode, with all extras enabled.